Complete the checkout function

+++ database/migrations/2019_01_26_154903_create_product_orders_table.php
@@ -15,7 +15,7 @@ class CreateProductOrdersTable extends Migration
   {
     Schema::create('product_orders', function (Blueprint $table) {
       $table->increments('id');
-      $table->unsignedInteger('order_id');
+      $table->unsignedBigInteger('order_id');
       $table->unsignedInteger('product_id');
       $table->unsignedInteger('quantity');
       $table->unsignedInteger('single_price');
